tx/ducktape: split upgrade tests into a separate test file
